+++ anaconda_opentelemetry/formatting.py
@@ -1,6 +1,5 @@
 from typing import Dict, Union, Sequence
 
-# Custom Typing
 Scalar = Union[str, bool, int, float]
 AttrDict = Dict[str, Union[str, bool, int, float, Sequence[Scalar]]]
 EventPayload = Union[str, Dict[Union[str, int, float, bool], Union[str, int, float, bool]]]